refactor

+++ shinkai-libs/shinkai-message-primitives/src/schemas/shinkai_subscription_req.rs
@@ -1,6 +1,7 @@
 use serde::{Deserialize, Serialize};
+use rust_decimal::Decimal;
 
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Eq)]
 pub struct FolderSubscription {
     pub minimum_token_delegation: Option<u64>,
     pub minimum_time_delegated_hours: Option<u64>,
@@ -10,10 +11,10 @@ pub struct FolderSubscription {
     pub folder_description: String,
 }
 
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Eq)]
 pub enum PaymentOption {
-    USD(f64),
-    KAITokens(u64),
+    USD(Decimal),
+    KAITokens(Decimal),
 }
 
 #[derive(Debug, Eq, Clone, PartialEq, Serialize, Deserialize)]
